% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_gn.R
\name{chi_prob}
\alias{chi_prob}
\title{chi_prob}
\usage{
chi_prob(alpha, n, c)
}
\arguments{
\item{alpha}{The significance level.}

\item{n}{The selected sample size.}

\item{c}{The value of c to use for this calculation.}
}
\value{
The calculated probabilities for the given parameters.
}
\description{
This function is used internally by several of the chapter 2 functions.
}
\examples{
chi_prob(0.05, 5, 1)

}
