% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_hn.R
\name{chi_prob_2}
\alias{chi_prob_2}
\title{chi_prob_2}
\usage{
chi_prob_2(alpha, n, c)
}
\arguments{
\item{alpha}{The selected significance level.}

\item{n}{The sample size.}

\item{c}{The c value.}
}
\value{
The calculated probability.
}
\description{
This function is called internally by \code{ProbHn_2}.
}
\examples{
chi_prob_2(0.05, 5, 1)

}
