% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_kn.R
\name{chi_prob_3}
\alias{chi_prob_3}
\title{chi_prob_3}
\usage{
chi_prob_3(alpha, n, c)
}
\arguments{
\item{alpha}{The selected significance level.}

\item{n}{The sample size.}

\item{c}{The c value.}
}
\value{
The calculated probability.
}
\description{
This function is called internally by \code{ProbKn_2}.
}
