% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_index.R
\name{gini_index}
\alias{gini_index}
\title{Gini Index}
\usage{
gini_index(x, na.rm = TRUE)
}
\arguments{
\item{x}{A vector containing the data for which the Gini index will be
calculated.}

\item{na.rm}{This parameter controls whether NA values are removed from
the vector (or not). Default is \code{TRUE}.}
}
\value{
The calculated value of the Gini index.
}
\description{
This function calculates the Gini Index for a vector \code{x}.
}
\examples{
# Household money from income brackets (2010) from Table C of
# https://en.wikipedia.org/wiki/Gini_coefficient. 
x <- c(13.7, 12.0, 10.9, 13.9, 17.7, 11.4, 12.1, 4.5, 3.9) 
gini_index(x)

}
