% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gini_mean_diff.R
\name{gini_mean_diff}
\alias{gini_mean_diff}
\title{Gini Mean Difference}
\usage{
gini_mean_diff(x, na.rm = TRUE)
}
\arguments{
\item{x}{The vector of values for which the Gini Mean Difference will
be calculated. This vector should be a numeric vector.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The calculated Gini Mean Difference.
}
\description{
This function calculates the Gini Mean Difference value. The input is
a vector of values for which the Gini Mean Difference is calculated.
}
\examples{
# Household money from income brackets (2010) from Table C of
# https://en.wikipedia.org/wiki/Gini_coefficient. 
x <- c(13.7, 12.0, 10.9, 13.9, 17.7, 11.4, 12.1, 4.5, 3.9)
gini_mean_diff(x)

}
