% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu13.R
\name{mu13}
\alias{mu13}
\title{mu13}
\usage{
mu13(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}

\item{na.rm}{This parameter controls whether NA values are removed from
the vector. Default is \code{FALSE}.}
}
\value{
The calculated value for use in other functions.
}
\description{
This an internal function called by other functions.
}
\examples{
x <- c(1:4)
y <- c(2:5)
mu13(x,y)

}
