% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu3.R
\name{mu3}
\alias{mu3}
\title{mu3}
\usage{
mu3(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector of at least length 3}

\item{na.rm}{logical value indicating whether NA values should be stripped
before the computation proceeds.}
}
\value{
The unbiased sample central moment of \code{X}
}
\description{
The function computes the unbiased estimate of the third central moment
}
\details{
The function \code{mu3} computes the unbiased estimate of the third central moment
of a given vector. The input must numeric (as determined by
\code{is.numeric}) and of length at least 3
(excluding \code{NA}'s. If \code{na.rm} is \code{TRUE}, the \code{NA}
entries in the given vector will be ignored.
}
\examples{
x <- rnorm(10)
mu3(x)
y <- c(1, 3, 1, 5, 6, 3, NA, 3 ,4, NA, 2)
mu3(y, na.rm = TRUE)

}
\author{
Ken Kelley \email{kkelley@nd.edu},
Francis Bilson Darku \email{fbilsond@nd.edu},
Bhargab Chattopadhyay \email{bhargab@utdallas.edu}
}
