% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu31.R
\name{mu31}
\alias{mu31}
\title{mu31}
\usage{
mu31(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{FALSE}.}
}
\value{
This function returns the calculated value for use in
other functions.
}
\description{
This an internal function called by other functions.
}
\examples{
x <- c(1:4)
y <- c(2:5)
mu31(x,y)

}
