% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu4.R
\name{mu4}
\alias{mu4}
\title{mu4}
\usage{
mu4(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of at least length 4.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{FALSE}.}
}
\value{
The calculated moment.
}
\description{
A function to compute the unbiased sample central fourth moment.
}
\details{
The function \code{mu4} computes the unbiased sample central moment
of a given vector. The input must numeric (as determined by
\code{is.numeric}) and of length at least 4
(excluding \code{NA}'s. If \code{na.rm} is \code{TRUE}, the \code{NA}
entries in the given vector will be ignored.
}
\examples{
x <- c(1:5)
mu4(x)

}
