% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_gn.R
\name{prob_gn}
\alias{prob_gn}
\title{ProbGn}
\usage{
prob_gn(alpha, n, c_min = 0, c_max = 10, by = 0.04, dec = 4)
}
\arguments{
\item{alpha}{The selected significance level.}

\item{n}{The selected sample size.}

\item{c_min}{The minimum value of c to try in the calulcations, default = 0.}

\item{c_max}{The maximum value of c to try in the calculations,
default = 10.}

\item{by}{The increment of the sequence between c_min and c_max,
default = 0.04.}

\item{dec}{The number of decimal places to use when removing probabilities
of 0 and 1 from the ends of the results.}
}
\value{
A sequence of values of c and the corresponding calculated
 probabilities.
}
\description{
This function replicates the ProbGn function from chapter 2.
This program computes the probability of K(n) > c*sigma for given alpha
and sample size.
}
\examples{
alpha = 0.05
n = 20
vals <- prob_gn(alpha = alpha, n = n)
vals

}
\references{
Mukhopadhyay, N., \& de Silva, B. M. (2009). \emph{Sequential Methods and Their Applications}. New York: CRC Press.
}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
