% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_kn.R
\name{prob_kn}
\alias{prob_kn}
\title{ProbKn}
\usage{
prob_kn(alpha, n, c_min = 0, c_max = 10, by = 0.04, dec = 4)
}
\arguments{
\item{alpha}{The significance level.}

\item{n}{The sample size.}

\item{c_min}{The minimum value of c to try in the calculations, default = 0.}

\item{c_max}{The maximum value of c to try in the calculations,
default = 10.}

\item{by}{The increment of the sequence between c_min and c_max,
default = 0.04.}

\item{dec}{The number of decimal places to use when removing probabilities
of 0 and 1 from the ends of the results.}
}
\value{
A sequence of values of c and the corresponding probabilities.
}
\description{
prob_kn computes the probability of K(n) > c*sigma for given alpha
and sample size.
}
\examples{
alpha = 0.05
n = 20
vals <- prob_kn(alpha = alpha, n = n)
vals

}
\references{
Mukhopadhyay, N., \& de Silva, B. M. (2009). \emph{Sequential Methods and Their Applications}. New York: CRC Press.
}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
