% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s.R
\name{s}
\alias{s}
\title{s}
\usage{
s(dat, i, j)
}
\arguments{
\item{dat}{A matrix with two columns}

\item{i}{power 1}

\item{j}{power 2}
}
\value{
The sum of \code{dat[,1]^i} and \code{dat[,2]^j}
}
\description{
This function is used internally to raise two vectors to the specified powers
and return the sum of their element-wise product.
}
\examples{
dat <- cbind(c(1:4),c(2:5))
s(dat,3,4)

}
