% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{sample_size}
\alias{sample_size}
\title{sample_size}
\usage{
sample_size(mu0, mu1, sigma, alpha, beta)
}
\arguments{
\item{mu0}{Mean value 0.}

\item{mu1}{Mean value 1.}

\item{sigma}{The standard deviation.}

\item{alpha}{The Type I error rate.}

\item{beta}{The desired Type II error rate, used when calculating the
necessary sample size.}
}
\value{
Returns the Type II error for a given sample size.
}
\description{
This function calculates the probability of Type II error for two independent groups when the 
when the population standard deviation is assumed known (i.e., using the $z$-distribution)#'  in order to achieve a preassigned Type II
error probability (beta) of interest.
}
