% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_aipe_contrast.R
\name{seq_aipe_contrast}
\alias{seq_aipe_contrast}
\title{seq_aipe_contrast}
\usage{
seq_aipe_contrast(
  alpha = 0.05,
  omega,
  data = NULL,
  coef = NULL,
  balanced = TRUE,
  Group.1 = NULL,
  Group.2 = NULL,
  pilot = FALSE,
  m0 = 4
)
}
\arguments{
\item{alpha}{The significance level., default is 0.05}

\item{omega}{omega,}

\item{data}{The dataset, default is NULL}

\item{coef}{The coefficients, default is NULL}

\item{balanced}{Is the data balanced, default is TRUE}

\item{Group.1}{The first group}

\item{Group.2}{The second group}

\item{pilot}{Is pilot of interest? Default is FALSE}

\item{m0}{The initial sample size.}
}
\value{
The current sample size, the current contrast, and an indicator of
if the criterion is satisfied.
}
\description{
Sequential approach to Accuracy in Parameter Estimation
for a Contrast
}
\examples{
pilot_ss <- seq_aipe_contrast(alpha=0.05, omega=0.2, pilot=TRUE)
SLS <- matrix(c(rexp(pilot_ss, rate=0.05),
rexp(pilot_ss, rate=0.05)), ncol = 2)
seq_aipe_contrast(alpha=0.05, omega=0.2,data = SLS)

}
\author{
Ken Kelley \email{KKelley@nd.edu},
Francis Bilson Darku \email{FBilsonD@nd.edu},
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}
}
