% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_aipe_cor.R
\name{seq_aipe_cor}
\alias{seq_aipe_cor}
\title{seq_aipe_cor}
\usage{
seq_aipe_cor(
  alpha = 0.05,
  omega,
  data = NULL,
  Group.1 = NULL,
  Group.2 = NULL,
  pilot = FALSE,
  m0 = 4,
  method = c("pearson", "kendall", "spearman"),
  na.rm = TRUE
)
}
\arguments{
\item{alpha}{The significance level., default is 0.05.}

\item{omega}{omega}

\item{data}{The data set, should have two columns.}

\item{Group.1}{The data for the first group.}

\item{Group.2}{The data for the second group.}

\item{pilot}{Should a pilot sample be generated.}

\item{m0}{The initial sample size.}

\item{method}{The correlation method to be used.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The current sample size, the current correlation and an
indicator of if the criterion has been satisfied.
}
\description{
Sequential approach to Accuracy in Parameter
Estimation for Effect Sizes
(AIPE): Correlation Coefficient
}
\examples{
pilot_ss <- seq_aipe_cor(alpha=0.05, omega=0.2, pilot=TRUE)
  SLS <- matrix(c(rexp(pilot_ss, rate=0.05),
  rexp(pilot_ss, rate=0.05)), ncol = 2)
seq_aipe_cor(alpha=0.05, omega=0.2,data = SLS)

}
\references{
Kelley, K., Darku, F. B., \& Chattopadhyay, B. (2018). Accuracy in parameter estimation for a general class of effect sizes: A sequential approach. \emph{Psychological Methods}, \emph{23}, 226–243.

Kelley, K., Darku, F. B., \& Chattopadhyay, B. (2019). Sequential Accuracy in parameter estimation for population correlation coefficients. \emph{Psychological Methods}, \emph{24}, 492–515.
}
\author{
Ken Kelley \email{KKelley@nd.edu},
Francis Bilson Darku \email{FBilsonD@nd.edu},
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}
}
