% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_aipe_cv.R
\name{seq_aipe_cv}
\alias{seq_aipe_cv}
\title{seq_aipe_cv}
\usage{
seq_aipe_cv(
  alpha = 0.05,
  omega,
  data = NULL,
  pilot = FALSE,
  m0 = 4,
  na.rm = TRUE
)
}
\arguments{
\item{alpha}{The significance level., default is 0.05.}

\item{omega}{omega}

\item{data}{The dataset for which to calculate the coefficient of
variation.}

\item{pilot}{Shoulda pilot sample be generated.}

\item{m0}{The initial sample size.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The current sample size, the coefficient of variation, an indicator
of if the criterion is satisfied, and the confidence interval.
}
\description{
Sequential approach to Accuracy in Parameter Estimation for Effect Sizes
(AIPE): Coefficient of Variation
}
\examples{
pilot_ss <- seq_aipe_cv(alpha=0.05, omega=0.2, pilot=TRUE)
SLS <- rexp(pilot_ss, rate=0.05)
seq_aipe_cv(alpha=0.05, omega=0.2,data = SLS)

}
\author{
Ken Kelley \email{KKelley@nd.edu},
Francis Bilson Darku \email{FBilsonD@nd.edu},
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}
}
