% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_aipe_gini.R
\name{seq_aipe_gini}
\alias{seq_aipe_gini}
\title{seq_aipe_gini}
\usage{
seq_aipe_gini(alpha = 0.05, omega, data = NULL, pilot = FALSE, m0 = 4)
}
\arguments{
\item{alpha}{he significance level., default is 0.05.}

\item{omega}{omega}

\item{data}{The data for which to calculate the gini index.}

\item{pilot}{Should a pilot sample be generated.}

\item{m0}{The initial sample size.}
}
\value{
The current sample size, the current gini coefficient, an
indicator of if the criterion is satisfied, and the confidence interval.
}
\description{
Sequential approach to Accuracy in Parameter
 Estimation for Effect Sizes
(AIPE): Gini Index
}
\examples{
pilot_ss <- seq_aipe_gini(alpha=0.05, omega=0.2, pilot=TRUE)
SLS <- rexp(pilot_ss, rate=0.05)
seq_aipe_gini(alpha=0.05, omega=0.2,data = SLS)

}
\author{
Ken Kelley \email{KKelley@nd.edu},
Francis Bilson Darku \email{FBilsonD@nd.edu},
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}
}
