% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_aipe_slr_slope.R
\name{seq_aipe_slr_slope}
\alias{seq_aipe_slr_slope}
\title{seq_aipe_slr_slope}
\usage{
seq_aipe_slr_slope(
  alpha = 0.05,
  omega,
  data = NULL,
  x = NULL,
  y = NULL,
  pilot = FALSE,
  m0 = 4,
  na.rm = TRUE
)
}
\arguments{
\item{alpha}{The significance level., default is 0.05.}

\item{omega}{omega}

\item{data}{The data for which to calculate the slope.}

\item{x}{The data vector for the independent variable.}

\item{y}{The data vector for the response.}

\item{pilot}{Should a pilot sample be generated.}

\item{m0}{The initial sample size.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The current sample size, the calculated slope, and an indicator of
if the criterion is satisfied.
}
\description{
Sequential approach to Accuracy in Parameter
 Estimation for Effect Sizes
(AIPE): Simple Linear Regression Slope
}
\examples{
pilot_ss <- seq_aipe_slr_slope(alpha=0.05, omega=0.2, pilot=TRUE)
SLS <- matrix( rnorm(pilot_ss[1],mean=0,sd=1),
 rnorm(pilot_ss[1], mean = 0, sd = 1), nrow=20, ncol= 2)
seq_aipe_slr_slope(alpha=0.05, omega=0.2,data = SLS)

}
\author{
Ken Kelley \email{KKelley@nd.edu},
Francis Bilson Darku \email{FBilsonD@nd.edu},
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}
}
