% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_aipe_smd.R
\name{seq_aipe_smd}
\alias{seq_aipe_smd}
\title{seq_aipe_smd}
\usage{
seq_aipe_smd(
  alpha = 0.05,
  omega,
  data = NULL,
  Group.1 = NULL,
  Group.2 = NULL,
  pilot = FALSE,
  m0 = 4,
  na.rm = TRUE
)
}
\arguments{
\item{alpha}{The significance level. default is 0.05.}

\item{omega}{omega}

\item{data}{The data set for which to calculate the standardized mean
difference.}

\item{Group.1}{The data vector for the first group.}

\item{Group.2}{The data vector for the second group.}

\item{pilot}{Should a pilot sample be generated.}

\item{m0}{The initial sample size.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The current sample size, the calculated standardized mean
difference, and an indicator of if the criterion has been satisfied.
}
\description{
Sequential approach to Accuracy in Parameter
Estimation for Effect Sizes
(AIPE): Standardized Mean Difference
}
\examples{
pilot_ss <- seq_aipe_smd(alpha=0.05, omega=0.2, pilot=TRUE)
SLS <- matrix( rnorm(pilot_ss[1],mean=0,sd=1),
rnorm(pilot_ss[1], mean = 0, sd = 1), nrow=20, ncol= 2)
seq_aipe_smd(alpha=0.05, omega=0.2,data = SLS)

}
\references{
Kelley, K., Darku, F. B., \& Chattopadhyay, B. (2018). Accuracy in parameter estimation for a general class of effect sizes: A sequential approach. \emph{Psychological Methods}, \emph{23}, 226–243.
}
\author{
Ken Kelley \email{KKelley@nd.edu},
Francis Bilson Darku \email{FBilsonD@nd.edu},
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}
}
