% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_blci_dfr_mean.R
\name{seq_blci_dfr_mean}
\alias{seq_blci_dfr_mean}
\title{seq_blci_dfr_mean}
\usage{
seq_blci_dfr_mean(data, d, alpha, pilot = FALSE, verbose = FALSE, na.rm = TRUE)
}
\arguments{
\item{data}{The data for which to calculate the confidence interval.
A numeric vector.}

\item{d}{Half of the confidence interval width, must be a non-zero positive
value.}

\item{alpha}{The significance level. A value between 0 and 1.}

\item{pilot}{Should a pilot sample be generated. True/False value.
default value is \code{FALSE}.}

\item{verbose}{Should the criterion be printed. Default is \code{FALSE}.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The calculated confidence interval, the sample size, data mean,
and an indicator of if the confidence level was satisfied.
}
\description{
Purely Sequential approach to Get Bounded Length Confidence Interval for the
median of distribution free random variables
}
\examples{
pilot_ss <- seq_blci_dfr_mean(alpha=0.05, d=1, pilot=TRUE)
SLS <- rnorm(pilot_ss, mean=0, sd=1)
seq_blci_dfr_mean(data=SLS, d=1, alpha=0.5, pilot=FALSE)

}
\references{
Mukhopadhyay, N., \& de Silva, B. M. (2009). \emph{Sequential Methods and Their Applications}. New York: CRC Press.
}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in},
Neetu Shah \email{201451015@iiitvadodara.ac.in}
}
