% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_br_norm_mean.R
\name{seq_br_norm_mean}
\alias{seq_br_norm_mean}
\title{seq_br_norm_mean}
\usage{
seq_br_norm_mean(
  data = NULL,
  A,
  k,
  w,
  pilot = FALSE,
  verbose = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{The data for which to calculate the bounded risk point.}

\item{A}{The loss function constant.}

\item{k}{Controls whether the absolute error or squared error is used.
K=1 uses absolute error, K=2 uses squared error.}

\item{w}{The risk bound.}

\item{pilot}{Should a pilot sample be generated. True/False value.
default value is \code{FALSE}.}

\item{verbose}{Should the criterion value be returned, True/False value,
default value is \code{FALSE}.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The calculated risk, the sample size, mean, standard deviation, and
an indicator of if the criterion was satisfied.
}
\description{
Sequential approach to Get Bounded Risk Point Estimation
for normal random variables. Calculates the risk for a normal
random variables.
}
\examples{
pilot_ss <- seq_br_norm_mean(A=2, k=1, w=0.4, pilot=TRUE)  # k=1 absolute error,
# k=2 squared error
SLS <- rnorm(pilot_ss, mean=2, sd=3)
seq_br_norm_mean(data=SLS, A=2, k=1, w=0.4, pilot=FALSE)
}
\references{
Mukhopadhyay, N., \& de Silva, Basil M. (2009). \emph{Sequential Methods and Their Applications}. New York: CRC Press.
}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in},
Neetu Shah \email{201451015@iiitvadodara.ac.in}
}
