% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_mr_dfr_mean_arb.R
\name{seq_mr_dfr_mean_arb}
\alias{seq_mr_dfr_mean_arb}
\title{seq_mr_dfr_mean_arb}
\usage{
seq_mr_dfr_mean_arb(
  data,
  A,
  c,
  gamma,
  pilot = FALSE,
  verbose = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{The data for which to calculate the confidence interval.
A numeric vector.}

\item{A}{The loss function constant.}

\item{c}{The cost of unit sample.}

\item{gamma}{gamma}

\item{pilot}{Should a pilot sample be generated. True/False value.
default value is \code{FALSE}.}

\item{verbose}{Should the criterion be printed. Default is \code{FALSE}.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The calculated risk, the sample size, mean, standard deviation,
 and
an indicator of if the criterion was satisfied.
}
\description{
Purely Sequential approach to Get Minimum Risk Point Estimation for the mean of distribution free random variables with arbitrary variance.
}
\examples{
pilot_ss <- seq_mr_dfr_mean_arb(A=100, c=2, pilot=TRUE)
SLS <- rnorm(pilot_ss, mean=0, sd=1)
seq_mr_dfr_mean_arb(data=SLS, A=100, c=2, gamma=0.5, pilot=FALSE)
}
\references{
Mukhopadhyay, N., \& de Silva, Basil M. (2009). \emph{Sequential Methods and Their Applications}. New York: CRC Press.
}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in},
Neetu Shah \email{201451015@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
