% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_mr_norm_mean.R
\name{seq_mr_norm_mean}
\alias{seq_mr_norm_mean}
\title{seq_mr_norm_mean}
\usage{
seq_mr_norm_mean(
  data,
  A,
  c,
  price,
  epsilon,
  pilot = FALSE,
  verbose = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{The data for which to calculate the minimum risk point.}

\item{A}{The loss function constant.}

\item{c}{The cost of a sampling a unit.}

\item{price}{price}

\item{epsilon}{epislon}

\item{pilot}{Should a pilot sample be generated. TRUE/FALSE value.
default value is \code{FALSE}.}

\item{verbose}{Should the criterion be printed. Default is \code{FALSE}.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The calculated risk, the sample size, mean, standard deviation, and
an indicator of if the criterion was satisfied.
}
\description{
Two Stage approach to Get minimum Risk Point Estimation
 for normal random variables.
}
\examples{
pilot_ss <- seq_mr_norm_mean(A=100, c=0.5, pilot=TRUE)
SLS <- rnorm(pilot_ss)
seq_mr_norm_mean(data=SLS, A=100, c=0.5, pilot=FALSE)
}
\references{
Mukhopadhyay, N., \& de Silva, Basil M. (2009). \emph{Sequential Methods and Their Applications}. New York: CRC Press.
}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in},
Neetu Shah \email{201451015@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
