% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_mr_smd.R
\name{seq_mr_smd}
\alias{seq_mr_smd}
\title{seq_mr_smd}
\usage{
seq_mr_smd(
  data1,
  data2,
  A,
  c1,
  c2,
  gamma,
  verbose = FALSE,
  pilot = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data1}{The first data vector for which to calculate the minimum
risk point.}

\item{data2}{The second data vector for which to calculate the minimum
risk point.}

\item{A}{The loss function constant.}

\item{c1}{The cost of unit sample for the first data vector.}

\item{c2}{The cost of unit sample for the second data vector.}

\item{gamma}{gamma}

\item{verbose}{Should the criterion be printed. Default is \code{FALSE}.}

\item{pilot}{Should a pilot sample be generated. TRUE/FALSE value.
default value is \code{FALSE}.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The calculated minimum risk point, the sample size of each data
vector, the mean of each vector, and an indicator of if the
criterion is satisfied.
}
\description{
Purely Sequential approach to Get Minimum Risk Point Estimation
 for the Standardized Mean Difference
}
\examples{
pilot_ss <- seq_mr_smd(gamma=1, A=100, c1=4, c2=5, pilot=TRUE)
SLS1 <- rnorm(pilot_ss[1], mean=0, sd=1)
SLS2 <- rnorm(pilot_ss[2], mean=0, sd=1)
seq_mr_smd(data1=SLS1, data2=SLS2, A=100, c1=4, c2=5, pilot=FALSE)
}
\references{
Chattopadhyay, B., & Kelley, K. (2017). Estimating the standardized mean
difference with minimum risk: Maximizing accuracy and minimizing
cost with sequential estimation. Psychological Methods, 22(1), 94-113
}
\author{
Ken Kelley \email{KKelley@nd.edu},
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in},
Neetu Shah \email{201451015@iiitvadodara.ac.in}

Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
