% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_ubdes_norm.R
\name{seq_ubdes_norm}
\alias{seq_ubdes_norm}
\title{seq_ubdes_norm}
\usage{
seq_ubdes_norm(
  data,
  C,
  K,
  w,
  alpha,
  pilot = FALSE,
  verbose = FALSE,
  na.rm = TRUE
)
}
\arguments{
\item{data}{The data matrix for which to calculate the}

\item{C}{A data frame.}

\item{K}{K}

\item{w}{The risk bound,A non-negative integer greater than zero.}

\item{alpha}{The significance level. A value between 0 and 1.}

\item{pilot}{Should a pilot sample be generated. True/False value.
default value is \code{FALSE}.}

\item{verbose}{Should the criterion be printed. Default is \code{FALSE}.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The calculated mean and an indicator of if the criterion is
satisfied.
}
\description{
Unbalanced Design using Purely Sequential approach
for two different size sets of normal random variables
}
\examples{
C <- matrix(c(2,3), nrow=2)
pilot_ss <- seq_ubdes_norm(alpha=0.05, w=4, C=C, pilot=TRUE)
SLS <- matrix( rnorm(pilot_ss[1],mean=0,sd=1), rnorm(pilot_ss[2],
mean = 0, sd = 1), nrow=2,ncol=2)
seq_ubdes_norm(data=SLS, C=C, 2, w=4, alpha=0.05, pilot=FALSE)

}
\references{
Ken Kelley, Francis Bilson Darku, and Bhargab Chattopadhyay. (2018).
Accuracy in Parameter Estimation for a General Class of
Effect Sizes: A Sequential Approach. \emph{Psychological Methods}, \emph{23}, 226--243.
}
\author{
Ken Kelley \email{KKelley@nd.edu},
Francis Bilson Darku \email{FBilsonD@nd.edu},
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in},
Neetu Shah \email{201451015@iiitvadodara.ac.in}
}
