% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seven_1.R
\name{seven_1}
\alias{seven_1}
\title{seven_1}
\usage{
seven_1(d, var, alpha)
}
\arguments{
\item{d}{Half the confidence interval width.}

\item{var}{The variance.}

\item{alpha}{The significance level.}
}
\value{
The calculate value of c.
}
\description{
This function replicates program 1 in seq 7 from the textbook.
}
\examples{
seven_1(d = 0.2, var = 10, alpha=0.01)

}
\references{
Mukhopadhyay, N., \& de Silva, Basil M. (2009). \emph{Sequential Methods and Their Applications}. New York: CRC Press.
}
