% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slr_slope.R
\name{slr_slope}
\alias{slr_slope}
\title{slr_slope}
\usage{
slr_slope(x, y, na.rm = TRUE)
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The calculated slope.
}
\description{
This function calcualtes the slope between two vectors.
}
\examples{
x <- c(1:4)
y <- c(2:5)
slr_slope(x, y)

}
