% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_br_exp_mean.R
\name{ts_br_exp_mean}
\alias{ts_br_exp_mean}
\title{ts_br_exp_mean}
\usage{
ts_br_exp_mean(data, A, w, pilot = FALSE, verbose = FALSE, na.rm = TRUE)
}
\arguments{
\item{data}{The data for which to calculate the bounded risk point.}

\item{A}{The loss function constant.}

\item{w}{The risk bound. A non-negative integer greater than zero.}

\item{pilot}{Should a pilot sample be generated. True/False value.
default value is \code{FALSE}.}

\item{verbose}{Should the criterion be printed. Default is \code{FALSE}.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{TRUE}.}
}
\value{
The current sample size, the mean and an indicator of if the
criterion is satisfied.
}
\description{
Two Stage approach to Get Bounded Risk Point
 Estimation for the mean of exponential random variables
}
\examples{
pilot_ss <- ts_br_exp_mean(w=4, pilot=TRUE)
SLS <- rexp(pilot_ss, rate=2)
\dontrun{
SLS <- rexp(pilot_ss, rate=2)
}
SLS <- rexp(100, rate=2)
ts_br_exp_mean(SLS, A=100, w=4, pilot=FALSE)
}
\references{
Mukhopadhyay, N., \& de Silva, Basil M. (2009). \emph{Sequential Methods and Their Applications}. New York: CRC Press.
}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in},
Neetu Shah \email{201451015@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
