% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{type_two_calc}
\alias{type_two_calc}
\title{type_two_calc}
\usage{
type_two_calc(mu0, mu1, sigma, alpha, n)
}
\arguments{
\item{mu0}{Mean value 0.}

\item{mu1}{Mean value 1.}

\item{sigma}{The standard deviation.}

\item{alpha}{The Type I error rate.}

\item{n}{The number of samples, used to calculate Type II error for a given
sample size.}
}
\value{
The Type II error for a given samples size.
}
\description{
This function is called internally by \code{power} and is used to
calculate the probability of Type II error.
}
