% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_contrast.R
\name{V2_contrast}
\alias{V2_contrast}
\title{V2_contrast}
\usage{
V2_contrast(data, coef, na.rm = FALSE)
}
\arguments{
\item{data}{The data set for which to calculate the contrast.}

\item{coef}{The set of coefficients.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{FALSE}.}
}
\value{
The coefficients multiplied by the variance of each column
}
\description{
This is an internal function called to calculate the estimate of the asymptotic variance of a contrast.
}
\examples{
dat_1 <- matrix(1:10, ncol=2)
V2_contrast(dat_1, coef=1)

}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
