% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_cor.R
\name{V2_cor}
\alias{V2_cor}
\title{V2_cor}
\usage{
V2_cor(x, y, method = c("pearson", "kendall", "spearman"), na.rm = FALSE)
}
\arguments{
\item{x}{The first vector.}

\item{y}{The second vector.}

\item{method}{The correlation method to be used.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{FALSE}.}
}
\value{
The estimated asymptotic variance of the calculated coefficient of variation.
}
\description{
This function is used internally to calculate the estimate of the asymptotic variance of the coefficient of variation.
}
\examples{
x <- c(1:4)
y <- c(2:5)
V2_cor(x,y, method = "pearson")


}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
