% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_cv.R
\name{V2_cv}
\alias{V2_cv}
\title{V2_cv}
\usage{
V2_cv(x, na.rm = FALSE)
}
\arguments{
\item{x}{The vector for which to calculate the coefficient of variation.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{FALSE}.}
}
\value{
The calculated coefficient of variation.
}
\description{
This function is used internally to calculate the coefficient of variation.
}
\examples{
x <- c(1:5)
V2_cv(x)

}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
