% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_gini_index.R
\name{V2_gini_index}
\alias{V2_gini_index}
\title{V2_gini_index}
\usage{
V2_gini_index(x, na.rm = FALSE)
}
\arguments{
\item{x}{The vector for which to calculate the Gini index.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{FALSE}.}
}
\value{
The estimated estimated asymptotic variance of the Gini index.
}
\description{
This function is used internally to calculate the Gini index.
}
\examples{
x <- c(1:4)
V2_gini_index(x)

}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
