% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_slr_slope.R
\name{V2_slr_slope}
\alias{V2_slr_slope}
\title{V2_slr_slope}
\usage{
V2_slr_slope(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{The independent variable data vector.}

\item{y}{The dependent variable data vector.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{FALSE}.}
}
\value{
The estimated asymptotic variance of the calculated slope.
}
\description{
This function is used internally to calculate the estimate of the asymptotic variance of the slope of two vectors.
}
\examples{
x <- c(1:4)
y <- c(2:5)
V2_slr_slope(x, y)

}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
