% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_smd.R
\name{V2_smd}
\alias{V2_smd}
\title{V2_smd}
\usage{
V2_smd(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{The first vector.}

\item{y}{The second vector.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{FALSE}.}
}
\value{
The estimated asymptotic variance of the standardized mean.
}
\description{
This function is used internally to calculate the estimate of the asymptotic variance of the standardized mean difference.
}
\examples{
x <- c(1:4)
y <- c(2:5)
V2_smd(x, y)

}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
