% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2_std_mean.R
\name{V2_std_mean}
\alias{V2_std_mean}
\title{V2_std_mean}
\usage{
V2_std_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{The vector for which to calculate the standardized mean.}

\item{na.rm}{This parameter controls whether NA values are removed from
the data prior to calculation. Default is \code{FALSE}.}
}
\value{
The estimated asymptotic variance of the standardized mean.
}
\description{
This function is used internally to calculate the estimate of the asymptotic variance of the standardized mean.
}
\examples{
x <- c(1:4)
V2_std_mean(x)

}
\references{
Kelley, K., Darku, F. B., & Chattopadhyay, B. (2018). Accuracy in parameter estimation for a general class of effect sizes: A sequential approach. \emph{Psychological Methods}, \emph{23}, 226–243.
}
\author{
Bhargab Chattopadhyay \email{Bhargab@iiitvadodara.ac.in}, Ken Kelley \email{kkelley@nd.edu}
}
